refactor(config): invert zeph-config dependency arrow and standardize provider fields
Closes #3481, Closes #3482

Moves pure-data config types from feature crates into zeph-config so that
the dependency arrow points in the correct direction (feature crates depend
on zeph-config, not the reverse). The following types are moved:
McpTrustLevel, ToolSecurityMeta, CacheTtl, ThinkingConfig, GeminiThinkingLevel,
ToolsConfig, AutonomyLevel, PreExecutionVerifierConfig, EvictionConfig,
CompressionGuidelinesConfig, CompactionProbeConfig, ProbeCategory.

Each source crate keeps a pub-use re-export for backward compatibility.
zeph-config/Cargo.toml no longer depends on zeph-llm, zeph-mcp, zeph-memory,
or zeph-tools.

Also standardizes five provider reference fields from raw String/Option<String>
to Option<ProviderName>: DigestConfig.provider, SemanticConfig.embed_provider,
HebbianConfig.consolidate_provider, GraphConfig.strategy_classifier_provider,
IndexConfig.embed_provider. Struct-level serde(default) ensures TOML backward
compatibility. Callers updated to use as_ref()/as_deref() patterns.
