+++ crates/zeph-config/Cargo.toml
@@ -19,6 +19,7 @@ default = []
 dirs.workspace = true
 ordered-float = { workspace = true, features = ["serde"] }
 regex.workspace = true
+schemars.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 thiserror.workspace = true
@@ -26,10 +27,6 @@ toml.workspace = true
 toml_edit.workspace = true
 tracing.workspace = true
 zeph-common.workspace = true
-zeph-memory.workspace = true
-zeph-mcp.workspace = true
-zeph-tools.workspace = true
-zeph-llm.workspace = true
 
 [dev-dependencies]
 insta.workspace = true
