+++ crates/zeph-llm/src/claude/types.rs
@@ -7,31 +7,8 @@ use serde::{Deserialize, Serialize};
 
 // ── Public types ──────────────────────────────────────────────────────────────
 
-/// Extended or adaptive thinking mode for Claude.
-///
-/// Serializes with `mode` as tag:
-/// `{ "mode": "extended", "budget_tokens": 10000 }` or `{ "mode": "adaptive" }`.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(tag = "mode", rename_all = "snake_case")]
-pub enum ThinkingConfig {
-    Extended {
-        budget_tokens: u32,
-    },
-    Adaptive {
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        effort: Option<ThinkingEffort>,
-    },
-}
-
-/// Effort level for adaptive thinking.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, Default, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-pub enum ThinkingEffort {
-    Low,
-    #[default]
-    Medium,
-    High,
-}
+// ThinkingConfig and ThinkingEffort are defined in zeph-config and re-exported here.
+pub use zeph_config::{ThinkingConfig, ThinkingEffort};
 
 pub(super) struct ThinkingCapability {
     /// Requires `interleaved-thinking-2025-05-14` beta header when `tool_use` is present.
@@ -383,40 +360,8 @@ pub(super) enum CacheType {
     Ephemeral,
 }
 
-/// Prompt cache TTL variant for the Claude API.
-///
-/// Controls how long a cached prompt prefix is retained. `Ephemeral` (default, ~5 min) is
-/// free to write; `OneHour` uses the `extended-cache-ttl-2025-04-25` beta and costs ~2× more
-/// to write but dramatically reduces re-write frequency for long-lived sessions.
-///
-/// # Serialization
-///
-/// When used as a TOML config value the accepted strings are `"ephemeral"` and `"1h"`.
-/// On the wire (Anthropic API), `OneHour` serializes as `"1h"` inside the `cache_control.ttl`
-/// field.
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum CacheTtl {
-    /// Default ephemeral TTL (~5 minutes). No beta header required.
-    #[default]
-    Ephemeral,
-    /// Extended 1-hour TTL. Requires the `extended-cache-ttl-2025-04-25` beta header.
-    /// Cache writes cost approximately 2× more than `Ephemeral`.
-    #[serde(rename = "1h")]
-    OneHour,
-}
-
-impl CacheTtl {
-    /// Returns `true` when this TTL variant requires the `extended-cache-ttl-2025-04-25` beta
-    /// header to be sent with each request.
-    #[must_use]
-    pub fn requires_beta(self) -> bool {
-        match self {
-            Self::OneHour => true,
-            Self::Ephemeral => false,
-        }
-    }
-}
+// CacheTtl is defined in zeph-config and re-exported here.
+pub use zeph_config::CacheTtl;
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub(super) struct CacheControl {
