+++ crates/zeph-mcp/Cargo.toml
@@ -40,6 +40,7 @@ tracing.workspace = true
 url.workspace = true
 uuid = { workspace = true, features = ["v5"] }
 zeph-common.workspace = true
+zeph-config.workspace = true
 zeph-llm.workspace = true
 zeph-memory = { workspace = true, features = ["sqlite"] }
 zeph-tools.workspace = true
